+++ api/src/backend/api/CHANGELOG.md
@@ -7,10 +7,8 @@ All notable changes to the **Prowler API** are documented in this file.
 
 ## [Unreleased]
 
-### Breaking Changes
-- Findings endpoints now require at least one date filter [(#6800)](https://github.com/prowler-cloud/prowler/pull/6800).
-
 ### Changed
 - Daily scheduled scan instances are now created beforehand with `SCHEDULED` state [(#6700)](https://github.com/prowler-cloud/prowler/pull/6700).
+- Findings endpoints now require at least one date filter [(#6800)](https://github.com/prowler-cloud/prowler/pull/6800).
 
 ---